Use situation optimized async scripts.
The original async script was optimized to run with the debug tool and
multiple files. But when no debug tool is available (= production mode)
or only one file is used, the async script was realy heavyweighted.
Now there are multiple implementations for single and multi file usage
and for production and development usage, that are optimized for the
specific situations. This make the async script more lightweight.